% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datasets.R
\name{descargar_datasets}
\alias{descargar_datasets}
\title{Descarga de datasets}
\usage{
descargar_datasets()
}
\description{
La función descargar_datasets() crea la carpeta datos en caso que no exista y descarga todos los archivos relacionados con la meteorología
}
\details{
!!CUIDADO, REESCRIBE LOS DATASETS EN CASO QUE YA EXISTAN!!
}
