% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datasets.R
\name{download_datasets}
\alias{download_datasets}
\title{Downloading datasets}
\usage{
download_datasets()
}
\description{
The function download_datasets() creates the 'data' folder if it does not already exist and downloads all files related to meteorology.
}
\details{
WARNING: IT WILL OVERWRITE EXISTING DATASETS!
}
