% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goatR-package.R
\docType{package}
\name{goatR-package}
\alias{goatR}
\alias{goatR-package}
\title{goatR: Geographical Observations and Analysis of Temperature and Rainfall}
\description{
Comprehensive toolkit for working with meteorological data from Argentine weather stations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bdirienzo.github.io/goatR/}
  \item Report bugs at \url{https://github.com/bdirienzo/goatR/issues}
}

}
\author{
\strong{Maintainer}: Bernardo Di Rienzo \email{bdirienzo@mail.austral.edu.ar}

Authors:
\itemize{
  \item Lautaro Sanfilippo \email{lsanfilippo@mail.austral.edu.ar}
}

}
\keyword{int?goaternal}
