% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_precipitation_plot.R
\name{monthly_precipitation_plot}
\alias{monthly_precipitation_plot}
\title{Monthly Precipitation Plot}
\usage{
monthly_precipitation_plot(
  ...,
  colors = NULL,
  title = "Accumulated Precipitation by Month"
)
}
\arguments{
\item{...}{One or more data frames, each containing data for a station. Each data frame must contain the columns 'fecha', 'id', and 'precipitacion_pluviometrica'.}

\item{colors}{A vector specifying the colors to use for the plot. If not provided, dark colors will be generated.}

\item{title}{The title of the plot. Defaults to "Accumulated Precipitation by Month" if not specified.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function generates a plot showing the total monthly precipitation for multiple stations.
}
\seealso{
\code{goatR::download_datasets} to download datasets from Argentine weather stations, and \code{goatR::read_datasets} to read the downloaded datasets into R.
}
