% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montly_temperature_plot.R
\name{monthly_temperature_plot}
\alias{monthly_temperature_plot}
\title{Generate a Monthly Temperature Plot}
\usage{
monthly_temperature_plot(..., colors = NULL, title = "Temperature")
}
\arguments{
\item{...}{One or more data frames, each containing data for a station. Each data frame must contain the columns 'fecha', 'id', and 'temperatura_abrigo_150cm'.}

\item{colors}{A vector specifying the colors to use for the plot. If not provided, dark colors will be generated.}

\item{title}{The title of the plot. Defaults to "Temperature" if not specified.}
}
\value{
A ggplot object representing the plot.
}
\description{
This function generates a plot showing the monthly average of 'temperatura_abrigo_150cm' for multiple stations.
}
