% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datasets.R
\name{read_datasets}
\alias{read_datasets}
\title{Read datasets}
\usage{
read_datasets()
}
\description{
The function read_datasets() reads all files related to meteorology from 'data_raw' folder and loads them into the global environment.
}
